Added CRUD APIs in the Callback Service and added respective test cases.
